Update to handling of html tags in wikicode
Added a global variable containing html tags which can be included in
wikicode and conserved when processed by the TinyMCE editor.  This
variable can be set in LocalSettings.php.  Also changed the behavior of
the context menu in TinyMCE to allow the user to select a meenu item to toggle it off so they
only see the browser context menu.

Change-Id: I916eaf59950148578ff15305785b778ea8c6d6fd