Fix infinite loop error
When two paragraphs with class mw_paragraph are followed by a list
or table element, this would lead to incorrect association of
opening and closing p tags. As a result, the parser would land in
an infinite loop. This commits uses a negative lookahead to make
sure only the paragraph directly preceeding a list or table is
affected.

Please note this error happens when one edits the main page in a
vanilla MediaWiki and might have some impact if not fixed.

Change-Id: I1bacdcda0e9b5a98352966c004206e9f7c61ad17