fixes for handling <pre> tags and tables
Fix for <pre> tags:  The way the TinyMCE editor currently handles <pre> tags is incompatible with how mediawiki handles them.  To avoid problems, this fix removes the TinyMCE menu dropdown for inserting <pre> blocks.  Mediawiki <pre> blocks can still be inserted using the '{;}' or '<>' menu items which allow input of wiki markup directly into the editor.  Future versions may address the incompatibility and re-enable this feature in the TinyMCE editor.
Fix for tables: There are three fixes here:
a) if there is a mediawiki table cell that contain only a '-' character, the space between the '|' and the '-' is removed when the TinyMCE editor converts the html content back to wiki markup.  This has now been corrected.
b) when table headers have no html attributes, the TinyMCE editor inserts a '|' after the '!' when converting the html back to wiki markup.  This has now been corrected.
c) the TinyMCE editor assumes that all tables have a class of 'wikitable'.  So all tables are saved with this class even if it wasn't present to start with.  This has now been corrected.  However, new tables inserted using the TinyMCE editor are still given a default class of 'wikitable'.  This can be changed on a case by case basis in TinyMCE when editing or the default can be over-ridden in the configuration file 'MW_tinymce.js'
Thank you to Viktor Schelling of Wikibase for spotting these issues.

Change-Id: I0fdde819fcb2560048734daa6fc58fbc95853d0a