Enabling SyntaxHighlight and fontawesome and other
Syntax SyntaxHighlight:
This change allows TinyMCE to add highlighting to text in
<syntaxhighlight> tags so it displays in the same way it would in the
wiki page.
fontawesome:
This change adds the fontawesome plugin for TinyMCE
(https://github.com/josh18/TinyMCE-FontAwesome-Plugin). There is a new
'flag' icon on the toolbar which allows the user to insert a fontawesome
icon into the page.  In order to display these icons on the wiki page
the following should be inserted at the top of the mediawiki:common.css
page in the wiki (without the line break after the 'font-')
@import url("extensions/TinyMCE/tinymce/plugins/fontawesome/css/font-
awesome.min.css");
There is a slight change to the original fontawesome plugin to allow the
'flag' icon to appear on the menu button for both small and large format
buttons.  The small button icons have also now been selected by default
in MW_tinymce.js
Other:
There have been some fixes to how new lines and horizontal rules are
handled.
There have also been improvements to how 'get' and 'set' events are
handled.
A different button icon is now used for the wikicode source editor
function which is '{{}}'.  This is in anticipation of introducing the
CodeMirror plugin in a furue release, which uses the '<>' icon.

Change-Id: I5a8c849ac150ceb1e6d38a4d4e2c9c8096bddd9b