images: fixes to dropping, pasting and uploading images
The TinyMCE extension had the ability to darg/drop or copy/paste html with images in such a way that the images would be uploaded to the wiki if they weren't already present.  Because of other fixes, this capability became broken and this patch fixes this.
Originally this functionality was in a TinyMCE plugin called 'mw_paste', which was a tailored version of the standard TinyMCE 'paste'.  This fix now uses the standard 'paste' plugin and implements the drop/paste upload behaviour in an event 'callback' in the main 'mw_wikicode' plugin.
At the same time, this fix also remedies the following related issues:
- a bug in the dropping/pasting of links
- a bug in the upload dialog that changed the format of images without a format to thumb.
- the treatment of non-editable wiki content was simplified and made more robust

Bug: T246826
Change-Id: I43286a3d34ddccf59cf6be1a815b544e6023ead9