Fix for problem causing slow load in some circumstances
When loading content into the editor, the extension checks whether
the content has already been parsed into html for TinyMCE, by checking
for the presence of elements with class names that start with 'mwt-'.
The code used in the match statement was very ineeficient and this has now
been improved. If your wikitext was long with a lot of classes declared then
it may have been very slow to load.  This should now not be the case.

Change-Id: I22bc3f05d9a195b2a9a4e1052a33f03477ae2a6d